Break Navigation into smaller parts
